Refactor: move time_state to engine/
