+++ openraft/src/engine/engine_impl.rs
@@ -11,6 +11,8 @@ use crate::engine::handler::replication_handler::SendNone;
 use crate::engine::handler::server_state_handler::ServerStateHandler;
 use crate::engine::handler::snapshot_handler::SnapshotHandler;
 use crate::engine::handler::vote_handler::VoteHandler;
+use crate::engine::time_state;
+use crate::engine::time_state::TimeState;
 use crate::engine::Command;
 use crate::entry::RaftEntry;
 use crate::error::ForwardToLeader;
@@ -25,8 +27,6 @@ use crate::raft::AppendEntriesResponse;
 use crate::raft::RaftRespTx;
 use crate::raft::VoteRequest;
 use crate::raft::VoteResponse;
-use crate::raft_state::time_state;
-use crate::raft_state::time_state::TimeState;
 use crate::raft_state::LogStateReader;
 use crate::raft_state::RaftState;
 use crate::summary::MessageSummary;
