+++ openraft/src/engine/handler/vote_handler.rs
@@ -1,12 +1,12 @@
 use crate::engine::engine_impl::EngineOutput;
 use crate::engine::handler::server_state_handler::ServerStateHandler;
+use crate::engine::time_state::TimeState;
 use crate::engine::Command;
 use crate::engine::EngineConfig;
 use crate::error::RejectVoteRequest;
 use crate::internal_server_state::InternalServerState;
 use crate::leader::Leader;
 use crate::progress::Progress;
-use crate::raft_state::time_state::TimeState;
 use crate::raft_state::LogStateReader;
 use crate::LogIdOptionExt;
 use crate::Node;
